requirements.txt
sh
Bio
edlib
primer3-py == 2.0.3
xmltodict
